+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0] - 2024-09-08
+
+### New Features
+
+* Handles Teledisk 1.x in addition to 2.x
+* Responds to a client's workspace symbols request
+* Make some use of backup FATs
+
+### Fixes
+
+* Formatting preserves blank lines
+* Correct a bug in rename symbol
+* Correct a bug in address hovers
+* Always zero high word of cluster1 for FAT12/16
+
 ## [3.0.2] - 2024-08-24
 
 ### Fixes